Wrap and stack annotations also in the UI endpoint for standalone sensor charts
